% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extract_pubmed.R
\name{.extract_fields}
\alias{.extract_fields}
\title{Extract fields of interest from a single PubMed article result file}
\usage{
.extract_fields(xml_doc)
}
\arguments{
\item{xml_doc}{The PubMed XML result file for a single study, as a string.}
}
\value{
A tibble of all articles of interest and their fields.
}
\description{
Extract all fields of interest for an article.
}
