% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extract_pubmed.R
\name{.extract_xml}
\alias{.extract_xml}
\title{Extract text of interest from an XML file}
\usage{
.extract_xml(xml_doc, xpath)
}
\arguments{
\item{xml_doc}{The XML document as it was returned from xml2.}
}
\value{
A string of interest.
}
\description{
Extract text of interest from an XML file.
}
