% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_unpaywall.R
\name{.get_unpaywall}
\alias{.get_unpaywall}
\title{Extract all information held about an article on Unpaywall.}
\usage{
.get_unpaywall(doi, email = "sstelios@gmail.com")
}
\arguments{
\item{email}{Your email address, as a string.}

\item{pubmed_xml}{A single DOI as a string.}
}
\value{
A tibble of all articles of interest and their fields.
}
\description{
Extract all information held about an article on Unpaywall.
}
