% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareadr.R
\docType{package}
\name{metareadr-package}
\alias{metareadr}
\alias{metareadr-package}
\title{metareadr: Download data for meta-research.}
\description{
It provides an interface to easily download data to facilitate meta-research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/serghiou/metareadr}
}

}
\author{
\strong{Maintainer}: Stylianos Serghiou \email{stelios.serghiou@gmail.com}

}
