% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_extract_pubmed.R
\name{mt_extract_pubmed}
\alias{mt_extract_pubmed}
\title{Extract fields of interest from PubMed results}
\usage{
mt_extract_pubmed(pubmed_xml, cl = 7)
}
\arguments{
\item{pubmed_xml}{The PubMed XML results file, as a string.}

\item{cl}{The number of CPU cores as a number to be passed to `pblapply`.}
}
\value{
A tibble of all articles of interest and their fields.
}
\description{
Extract all fields of interest for each article.
}
