% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_match.R
\name{mt_match}
\alias{mt_match}
\title{Find an issue-match to a PubMed article}
\usage{
mt_match(pmid = 16060722, n_max = 100)
}
\arguments{
\item{pmid}{The PMID of the article to which matches are desired.}

\item{n_max}{The maximum number of records to download from the issue.}
}
\value{
A table of articles from the same journal issue as the provided PMID.
}
\description{
Identify the PMID of an article in the same issue as the provided article.
}
