% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pdf_link.R
\name{mt_pdf_link}
\alias{mt_pdf_link}
\title{Extract PDF links from CrossRef}
\usage{
mt_pdf_link(doi)
}
\arguments{
\item{doi}{The DOI of an article, as a string.}
}
\value{
The URL to the PDF of an article, as a string.
}
\description{
Get the link to the PDF of an article, given its DOI.
}
