% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_read_pmcoa.R
\name{mt_read_pmcoa}
\alias{mt_read_pmcoa}
\title{Download file from PubMed Central}
\usage{
mt_read_pmcoa(pmcid, file_name = NULL, file_format = "pmc")
}
\arguments{
\item{pmcid}{The PMCID (not PMID) of the article as a string or integer
(e.g. 13900).}

\item{file_name}{The file name, as a string (e.g. "my_folder/my_file.xml").
If NULL (default), it will save the file in the current directory under
its PMCID (e.g. "PMC 13900").}

\item{file_format}{Format as described on the OAI-PMH service. The default
("pmc") downloads the whole text in XML format.}
}
\value{
Save the PubMed result as an XML file.
}
\description{
Download the file of the designated PMCID (PubMed Central ID) in the file
    type of interest. This uses the
    [OAI-PMH service](https://www.ncbi.nlm.nih.gov/pmc/tools/oai/).
    The Open Access Index only maintains the full text for a subset of the
    articles on PMC (not sure which subset).
}
