% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_sample_pmc.R
\name{mt_sample_pmc}
\alias{mt_sample_pmc}
\title{A random sample from PubMed Central (PMC)}
\usage{
mt_sample_pmc(n = 1000, year_min = 2016, year_max = 2020, seed = NULL)
}
\arguments{
\item{n}{The number of articles to sample, as a numeric.}

\item{year_min}{The minimum year to sample from inclusive, as a numeric.}

\item{year_max}{The maximum year to sample from inclusive, as a numeric.}

\item{seed}{Provide a seed for a reproducible sample, as a numeric.}
}
\value{
A tibble of randomly sampled articles and their PMCID, PMID, DOI,
    Journal Title, ISSN, eISSN, Year, Volume, Issue, Page and Date sampled.
    Date sampled is in yyyy-mm-dd.
}
\description{
Identify a random sample of articles. Note that this function retrieves the
    list of currently available articles from PMC, so it requires internet
    connection to download a file of ~550 MB at the time of writing and,
    depending on the connection, it will take about a minute to run. Note
    that these samples will change with changes in the data provided by PMC,
    for which reason a "Data sampled" column is added as a time-stamp.
}
