% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_sample_pubmed.R
\name{mt_sample_pubmed}
\alias{mt_sample_pubmed}
\title{A random sample from PubMed}
\usage{
mt_sample_pubmed(n = 1000, year_min = 2020, year_max = 2020, seed = NULL)
}
\arguments{
\item{n}{The number of articles to sample, as a numeric.}

\item{year_min}{The minimum year to sample from inclusive, as a numeric. By
PubMed definition, this is the year of record creation.}

\item{year_max}{The maximum year to sample from inclusive, as a numeric. By
PubMed definition, this is the year of record creation.}

\item{seed}{Provide a seed for a reproducible sample, as a numeric.}
}
\value{
A tibble of randomly sampled articles and their PMID and Date sampled
    in yyyy-mm-dd.
}
\description{
Identify a random sample of articles from PubMed. Note that this function
    retrieves the currently available articles from PubMed, so it requires
    internet connection to search PubMed and download the results. Depending
    on the connection and selected years, this can take at least a few
    minutes to run. Note that these samples will change with changes in the
    data provided by PubMed, for which reason a "Data sampled" column is
    added as a time-stamp. Please also consider using a seed to improve
    reproducibility and do not run this function repeatedly.
}
