% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_scihub.R
\name{mt_scihub}
\alias{mt_scihub}
\title{Extract PDFs from SciHub}
\usage{
mt_scihub(x, filename = NULL, filedir = "PDF_scihub")
}
\arguments{
\item{x}{Either a DOI or PMID as a string.}

\item{filename}{The name of the file to be saved, as a string.}

\item{filedir}{The name of the directory in which to save the file.}
}
\value{
Saves the PDF in the specified directory with the specified name.
}
\description{
Extract PDFs from SciHub. Note that this may be illegal in your country or
    institution. Note, also, that SciHub is very strict with automated
    downloads and using a VPN to cycle between IPs may be necessary.
}
