% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_read_pubmed.R
\name{mt_read_pubmed}
\alias{mt_read_pubmed}
\title{Download PubMed results as XML}
\usage{
mt_read_pubmed(term, n_max = 10000)
}
\arguments{
\item{term}{The search term, e.g. "meta-analysis[TITLE]", as a string.}

\item{n_max}{The maximum number of records to download, as an integer.}
}
\value{
Save the PubMed result as an XML file.
}
\description{
Saves the meta-data of a PubMed query as an XML file.
}
